pyproject.toml
[tool.poetry]
name = "nova-env"
version = "0.1.0"
description = ""
authors = ["gokberk <mgokberkoguz@gmail.com>"]
readme = "README.md"
packages = [{include = "nova_env"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
django-rest-framework = "^0.1.0"
markdown = "^3.4.3"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
